+++ src/views/User.vue
@@ -7,8 +7,7 @@
       <li :key="item.key" v-for="item in knowledges">{{item}}</li>
     </ul>
 
-    <input v-model="form.title" class="input" type="text" placeholder="Title">
-    <button @click="addKnowledge">add</button>
+    <KnowledgeForm/>
   </div>
 </template>
 
@@ -18,44 +17,24 @@ import { Component, Vue } from "vue-property-decorator";
 
 import { Knowledge } from "../domains/knowledges/model";
 import { config } from "../domains/knowledges/config";
+import KnowledgeForm from "@/components/KnowledgeForm";
 
 @Component({
-  components: {},
+  components: {
+    KnowledgeForm
+  },
   data() {
     const user = firebase.auth().currentUser;
 
     return {
       displayName: user ? user.displayName : null,
-      knowledges: [],
-      form: {
-        title: ""
-      }
+      knowledges: []
     };
   },
   mounted() {
     this.getKnowledges();
   },
   methods: {
-    addKnowledge: async function() {
-      const { title } = this.$data.form;
-      console.log("this", title);
-
-      const currentUser = firebase.auth().currentUser;
-      const uid = currentUser.uid;
-
-      const collection = firebase
-        .firestore()
-        .collection(config.collection_endpoint);
-
-      await collection.add({
-        title,
-        author: currentUser.uid
-      });
-
-      await this.getKnowledges();
-
-      console.log("currentUser", uid);
-    },
     logout: function() {
       firebase.auth().signOut();
       this.$router.replace("home");
